requirements.txt
streamlit==1.30.0
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7
html2text==2020.1.16
reportlab==4.0.4
google-generativeai==0.3.2
Pillow>=10.1.0
python-dotenv>=1.0.0